+++ worker/src/worker.ts
@@ -190,6 +190,7 @@ app.use('/user_api/*', async (c, next) => {
 		|| c.req.path.startsWith("/user_api/verify_code")
 		|| c.req.path.startsWith("/user_api/passkey/authenticate_")
 		|| c.req.path.startsWith("/user_api/oauth2")
+		|| c.req.path.startsWith("/user_api/oidc")
 	) {
 		await next();
 		return;